requirements.txt
python_Levenshtein==0.12.0
dill==0.3.0
GitPython==3.0.2
imohash==1.0.4
joblib==0.13.2
nevergrad==0.2.3
nltk==3.4.3
numpy==1.17.2
pandas==0.25.1
sentencepiece==0.1.83
spacy==2.1.3
tabulate==0.8.4
torch==1.2.0
tqdm==4.36.1
easse@git+git://github.com/feralvam/easse.git@5dce4474a72baa5a16e3764f1ed4225a1751dbf2
fairseq@git+https://github.com/louismartin/fairseq.git@controllable-sentence-simplification
